# www/public/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "description": "Issued when a browser pairs with the daemon.",
        "in": "cookie",
        "name": "picode_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP API the PiCode web UI itself speaks — same daemon, same pairing-gated session. Generated from the server's route registration by cmd/picode-openapi; do not edit the JSON by hand. Authenticate with the picode_session cookie issued when you pair a browser (Settings → Pairing).",
    "title": "PiCode HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/terminals": {
      "get": {
        "operationId": "getApiTerminals",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      },
      "post": {
        "operationId": "postApiTerminals",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
...
    "/api/terminals/{id}": {
      "delete": {
        "operationId": "deleteApiTerminals",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      },
      "patch": {
        "operationId": "patchApiTerminals",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
    "/api/terminals/{id}/blob": {
      "get": {
        "operationId": "getApiTerminalsBlob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
    "/api/terminals/{id}/browse": {
      "get": {
        "operationId": "getApiTerminalsBrowse",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
    "/api/terminals/{id}/cwd": {
      "get": {
        "operationId": "getApiTerminalsCwd",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
...
    "/api/terminals/{id}/open": {
      "post": {
        "operationId": "postApiTerminalsOpen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
...
    "/api/terminals/{id}/state": {
      "post": {
        "operationId": "postApiTerminalsState",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
    "/api/terminals/{id}/text": {
      "get": {
        "operationId": "getApiTerminalsText",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      },
      "put": {
        "operationId": "putApiTerminalsText",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "terminals"
        ]
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "local daemon",
      "url": "http://127.0.0.1:8445"
    }
  ],
...
}